Remove all code for generic xattr support.
It doesn't seem likely that xattrs can become a viable feature on their own because they generally need some interpretation to be useful (e.g. SELinux contexts, POSIX ACLs).

So remove all that code and structure the code more like the SELinux code was implemented as an extension. We may not do that, but at least it makes the patch much easier to maintain.